# apps/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  entity.UserNutritionGoal:
    properties:
      created_at:
        type: string
      goal_effective_date:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      notes:
        type: string
      target_calories:
        type: number
      target_carbs_grams:
        type: number
      target_fat_grams:
        type: number
      target_fiber_grams:
        type: number
      target_protein_grams:
        type: number
      target_sugar_grams_limit:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  v1.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Fitness application API for tracking nutrition, meals, and products
  title: Rebound API
  version: "1.0"
paths:
...
  /nutrition/goals:
    post:
      consumes:
      - application/json
      description: Create new nutrition goals for a user
      parameters:
      - description: Nutrition goals object
        in: body
        name: goals
        required: true
        schema:
          $ref: '#/definitions/entity.UserNutritionGoal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.UserNutritionGoal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Create nutrition goals
      tags:
      - nutrition
  /nutrition/goals/{id}:
    delete:
      consumes:
      - application/json
      description: Delete nutrition goals by ID
      parameters:
      - description: Nutrition goals ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete nutrition goals
      tags:
      - nutrition
    get:
      consumes:
      - application/json
      description: Get nutrition goals by ID
      parameters:
      - description: Nutrition goals ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserNutritionGoal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get nutrition goals
      tags:
      - nutrition
    put:
      consumes:
      - application/json
      description: Update existing nutrition goals
      parameters:
      - description: Nutrition goals ID
        in: path
        name: id
        required: true
        type: string
      - description: Nutrition goals object
        in: body
        name: goals
        required: true
        schema:
          $ref: '#/definitions/entity.UserNutritionGoal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserNutritionGoal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update nutrition goals
      tags:
      - nutrition
  /nutrition/goals/user/{userID}/active:
    get:
      consumes:
      - application/json
      description: Get the active nutrition goals for a user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserNutritionGoal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get active nutrition goals
      tags:
      - nutrition
  /nutrition/goals/user/{userID}/history:
    get:
      consumes:
      - application/json
      description: Get nutrition goals history for a user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.UserNutritionGoal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get nutrition goals history
      tags:
      - nutrition
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"